src/requirements.txt
# Image Encoder Dependencies
Pillow>=9.0.0
numpy>=1.20.0
